Model/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docopt==0.6.2
env==0.1.0
filelock==3.12.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
imageio==2.29.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lazy_loader==0.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pipreqs==0.4.13
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
SQLAlchemy==2.0.15
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.2
urllib3==2.0.2
Werkzeug==2.3.4
yarg==0.1.9
zipp==3.15.0
gunicorn
python-dotenv
apscheduler